plugins/gvrelgraph: create family nodes for fam. with only children
See #0010156 -- If a tree contains a family with more than 1 child, but
no parents, the graph would not contain a family node for it, but create
two detached sub-trees.

TODO: find a way to attach the siblings if family nodes are disabled
(i.e. they would be directly linked to their parents -- which we don't
have here)